Maintenance update

+++ tests/conftest.py
@@ -1,3 +1,5 @@
+"""Test configuration and fixtures."""
+
 import logging
 import os
 from pathlib import Path
@@ -6,20 +8,20 @@ from shutil import rmtree
 import pytest
 
 
-@pytest.fixture(scope="package", autouse=True)
+@pytest.fixture(scope="session", autouse=True)
 def chdir() -> None:
     """
     Fixture that changes the current working directory to the 'test_working_directory' folder.
-    This fixture is automatically used for the entire package.
+    This fixture is automatically used for the entire session.
     """
     os.chdir(Path(__file__).parent.absolute() / "test_working_directory")
 
 
-@pytest.fixture(scope="package", autouse=True)
+@pytest.fixture(scope="session", autouse=True)
 def test_dir() -> Path:
     """
     Fixture that returns the absolute path of the directory containing the current file.
-    This fixture is automatically used for the entire package.
+    This fixture is automatically used for the entire session.
     """
     return Path(__file__).parent.absolute()
 
@@ -29,7 +31,12 @@ output_dirs = [
 ]
 output_files = [
     "*test*.pdf",
-    "systemModel.xml",
+    "*.xml",
+    "*.log",
+    "*.js5",
+    "*.cases",
+    "test_results",
+    "*.fmu",
 ]
 
 
@@ -53,7 +60,8 @@ def _remove_output_dirs_and_files() -> None:
     for pattern in output_files:
         for file in Path.cwd().glob(pattern):
             _file = Path(file)
-            _file.unlink(missing_ok=True)
+            if _file.is_file():
+                _file.unlink(missing_ok=True)
 
 
 @pytest.fixture(autouse=True)
@@ -73,9 +81,14 @@ def logger() -> logging.Logger:
 
 
 def pytest_addoption(parser: pytest.Parser) -> None:
-    parser.addoption("--show", action="store", default=False)
+    parser.addoption(
+        "--show",
+        action="store_true",
+        default=False,
+        help="Command line switch to show plots during tests, and dump additional results to console. By default, False.",
+    )
 
 
 @pytest.fixture(scope="session")
 def show(request: pytest.FixtureRequest) -> bool:
-    return request.config.getoption("--show") == "True"
+    return request.config.getoption("--show")
